+++ lib/screens/chapterSelection.dart
@@ -1,7 +1,7 @@
 import 'package:fuenfzigohm/constants.dart';
-import 'package:fuenfzigohm/coustom_libs/database.dart';
-import 'package:fuenfzigohm/coustom_libs/icons.dart';
-import 'package:fuenfzigohm/coustom_libs/json.dart';
+import 'package:fuenfzigohm/custom_libs/database.dart';
+import 'package:fuenfzigohm/custom_libs/icons.dart';
+import 'package:fuenfzigohm/custom_libs/json.dart';
 import 'package:fuenfzigohm/screens/question.dart';
 import 'package:fuenfzigohm/screens/settings.dart';
 import 'package:fuenfzigohm/screens/aboutApp.dart';
