+++ lib/screens/intro.dart
@@ -1,4 +1,4 @@
-import 'package:fuenfzigohm/coustom_libs/database.dart';
+import 'package:fuenfzigohm/custom_libs/database.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:fuenfzigohm/screens/chapterSelection.dart';
