+++ lib/screens/selectLearningPath.dart
@@ -1,7 +1,7 @@
 
 
 
-import 'package:fuenfzigohm/coustom_libs/database.dart';
+import 'package:fuenfzigohm/custom_libs/database.dart';
 import 'package:flutter/material.dart';
 
 class selectClass extends StatelessWidget{
