+++ lib/screens/settings.dart
@@ -1,4 +1,4 @@
-import 'package:fuenfzigohm/coustom_libs/database.dart';
+import 'package:fuenfzigohm/custom_libs/database.dart';
 import 'package:flutter/material.dart';
 import 'package:settings_ui/settings_ui.dart';
